+++ lizard_wbconfiguration/templates/lizard_wbconfiguration/wbconfiguration_read_only.js
@@ -2,6 +2,10 @@
 {% load get_grid %}
 {% load get_portal_template %}
 {
-    xtype: 'wb_grid_history'
+    xtype: 'wb_grid_history',
+    config: {
+        log_entry_id: {{ view.log_entry_id }}
 
+    },
+    log_entry_id: {{ view.log_entry_id }}
 }